tmalign: another, more thorough deduplication option
For each duplicated sequence (potentially aligned differently), keep only the
copy with the best TM-score, as stored in the FASTA header.

For fammer, this actually doesn't perform as well as the version that keeps
multiple, differently aligned copies of the same sequence, but maybe it will be
handy for stand-alone usage. (Not enabled through a command-line option yet.)